Add linenoise
Poor little ed really doesn't want this library to be forced in...
too bad! Lots of hacky stuff going on here, and it was a PITA to debug,
but it (kinda) works (at least from what I can see). The best way to
test this would be to dogfood it, so that's exactly what I'll do.